LazyColumn
Learning how to use a LazyColumn in Jetpack Compose for displaying large lists in the app. Adding and developing a model named 'Person' and a repository containing a list of persons. Customizing list items and creating a LazyColumn to display all data in the items. This includes default LazyColumn with items, items indexed, and a LazyColumn using sticky headers.